Add KeepOpenFile, LazyFile and open_file files.
