+++ tests/test_utils.py
@@ -1,145 +1,10 @@
 import os
 import pathlib
-import stat
 
 import pytest
 
 import click._termui_impl
 import click.utils
-from click._compat import WIN
-
-
-def test_open_file(runner):
-    @click.command()
-    @click.argument("filename")
-    def cli(filename):
-        with click.open_file(filename) as f:
-            click.echo(f.read())
-
-        click.echo("meep")
-
-    with runner.isolated_filesystem():
-        with open("hello.txt", "w") as f:
-            f.write("Cool stuff")
-
-        result = runner.invoke(cli, ["hello.txt"])
-        assert result.exception is None
-        assert result.output == "Cool stuff\nmeep\n"
-
-        result = runner.invoke(cli, ["-"], input="foobar")
-        assert result.exception is None
-        assert result.output == "foobar\nmeep\n"
-
-
-def test_open_file_pathlib_dash(runner):
-    @click.command()
-    @click.argument(
-        "filename", type=click.Path(allow_dash=True, path_type=pathlib.Path)
-    )
-    def cli(filename):
-        click.echo(str(type(filename)))
-
-        with click.open_file(filename) as f:
-            click.echo(f.read())
-
-        result = runner.invoke(cli, ["-"], input="value")
-        assert result.exception is None
-        assert result.output == "pathlib.Path\nvalue\n"
-
-
-def test_open_file_ignore_errors_stdin(runner):
-    @click.command()
-    @click.argument("filename")
-    def cli(filename):
-        with click.open_file(filename, errors="ignore") as f:
-            click.echo(f.read())
-
-    result = runner.invoke(cli, ["-"], input=os.urandom(16))
-    assert result.exception is None
-
-
-def test_open_file_respects_ignore(runner):
-    with runner.isolated_filesystem():
-        with open("test.txt", "w") as f:
-            f.write("Hello world!")
-
-        with click.open_file("test.txt", encoding="utf8", errors="ignore") as f:
-            assert f.errors == "ignore"
-
-
-def test_open_file_ignore_invalid_utf8(runner):
-    with runner.isolated_filesystem():
-        with open("test.txt", "wb") as f:
-            f.write(b"\xe2\x28\xa1")
-
-        with click.open_file("test.txt", encoding="utf8", errors="ignore") as f:
-            f.read()
-
-
-def test_open_file_ignore_no_encoding(runner):
-    with runner.isolated_filesystem():
-        with open("test.bin", "wb") as f:
-            f.write(os.urandom(16))
-
-        with click.open_file("test.bin", errors="ignore") as f:
-            f.read()
-
-
-@pytest.mark.skipif(WIN, reason="os.chmod() is not fully supported on Windows.")
-@pytest.mark.parametrize("permissions", [0o400, 0o444, 0o600, 0o644])
-def test_open_file_atomic_permissions_existing_file(runner, permissions):
-    with runner.isolated_filesystem():
-        with open("existing.txt", "w") as f:
-            f.write("content")
-        os.chmod("existing.txt", permissions)
-
-        @click.command()
-        @click.argument("filename")
-        def cli(filename):
-            click.open_file(filename, "w", atomic=True).close()
-
-        result = runner.invoke(cli, ["existing.txt"])
-        assert result.exception is None
-        assert stat.S_IMODE(os.stat("existing.txt").st_mode) == permissions
-
-
-@pytest.mark.skipif(WIN, reason="os.stat() is not fully supported on Windows.")
-def test_open_file_atomic_permissions_new_file(runner):
-    with runner.isolated_filesystem():
-
-        @click.command()
-        @click.argument("filename")
-        def cli(filename):
-            click.open_file(filename, "w", atomic=True).close()
-
-        # Create a test file to get the expected permissions for new files
-        # according to the current umask.
-        with open("test.txt", "w"):
-            pass
-        permissions = stat.S_IMODE(os.stat("test.txt").st_mode)
-
-        result = runner.invoke(cli, ["new.txt"])
-        assert result.exception is None
-        assert stat.S_IMODE(os.stat("new.txt").st_mode) == permissions
-
-
-def test_iter_keepopenfile(tmpdir):
-    expected = list(map(str, range(10)))
-    p = tmpdir.mkdir("testdir").join("testfile")
-    p.write("\n".join(expected))
-    with p.open() as f:
-        for e_line, a_line in zip(expected, click.utils.KeepOpenFile(f), strict=False):
-            assert e_line == a_line.strip()
-
-
-def test_iter_lazyfile(tmpdir):
-    expected = list(map(str, range(10)))
-    p = tmpdir.mkdir("testdir").join("testfile")
-    p.write("\n".join(expected))
-    with p.open() as f:
-        with click.utils.LazyFile(f.name) as lf:
-            for e_line, a_line in zip(expected, lf, strict=False):
-                assert e_line == a_line.strip()
 
 
 class MockMain:
