+++ tests/test_utils_2/test_KeepOpenFile.py
@@ -1,0 +1,10 @@
+import click
+
+
+def test_iter_keepopenfile(tmpdir):
+    expected = list(map(str, range(10)))
+    p = tmpdir.mkdir("testdir").join("testfile")
+    p.write("\n".join(expected))
+    with p.open() as f:
+        for e_line, a_line in zip(expected, click.utils.KeepOpenFile(f), strict=False):
+            assert e_line == a_line.strip()
