+++ tests/test_utils_2/test_LazyFile.py
@@ -1,0 +1,11 @@
+import click
+
+
+def test_iter_lazyfile(tmpdir):
+    expected = list(map(str, range(10)))
+    p = tmpdir.mkdir("testdir").join("testfile")
+    p.write("\n".join(expected))
+    with p.open() as f:
+        with click.utils.LazyFile(f.name) as lf:
+            for e_line, a_line in zip(expected, lf, strict=False):
+                assert e_line == a_line.strip()
